requirements.txt
asgiref==3.11.0
bleach==6.1.0
dj-database-url==3.0.1
Django==6.0
django-taggit==6.1.0
gunicorn==23.0.0
Markdown==3.10
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.8.2

